[#1447] reorganize classes
