+++ app/src/main/java/org/akvo/flow/presentation/geoshape/ViewGeoShapeActivity.java
@@ -31,8 +31,8 @@ import org.akvo.flow.activity.BackActivity;
 import org.akvo.flow.injector.component.DaggerViewComponent;
 import org.akvo.flow.injector.component.ViewComponent;
 import org.akvo.flow.offlinemaps.presentation.geoshapes.GeoShapesMapViewImpl;
-import org.akvo.flow.presentation.geoshape.create.FeatureMapper;
-import org.akvo.flow.presentation.geoshape.create.ViewFeatures;
+import org.akvo.flow.presentation.geoshape.create.entities.FeatureMapper;
+import org.akvo.flow.presentation.geoshape.create.entities.ViewFeatures;
 import org.akvo.flow.presentation.geoshape.create.entities.Shape;
 import org.akvo.flow.util.ConstantUtil;
