+++ app/src/main/java/org/akvo/flow/presentation/geoshape/create/entities/CoordinatesMapper.java
@@ -17,7 +17,7 @@
  * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.akvo.flow.presentation.geoshape.create;
+package org.akvo.flow.presentation.geoshape.create.entities;
 
 import com.mapbox.geojson.Point;
 import com.mapbox.mapboxsdk.geometry.LatLng;
