+++ app/src/main/java/org/akvo/flow/presentation/geoshape/create/entities/FeatureMapper.java
@@ -17,7 +17,7 @@
  * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.akvo.flow.presentation.geoshape.create;
+package org.akvo.flow.presentation.geoshape.create.entities;
 
 import com.mapbox.geojson.Feature;
 import com.mapbox.geojson.FeatureCollection;
@@ -30,11 +30,13 @@ import com.mapbox.mapboxsdk.geometry.LatLng;
 
 import org.akvo.flow.offlinemaps.presentation.geoshapes.GeoShapeConstants;
 import org.akvo.flow.presentation.geoshape.create.entities.AreaShape;
+import org.akvo.flow.presentation.geoshape.create.entities.CoordinatesMapper;
 import org.akvo.flow.presentation.geoshape.create.entities.LineShape;
 import org.akvo.flow.presentation.geoshape.create.entities.PointShape;
 import org.akvo.flow.presentation.geoshape.create.entities.PointsLatLngMapper;
 import org.akvo.flow.presentation.geoshape.create.entities.Shape;
 import org.akvo.flow.presentation.geoshape.create.entities.ShapePoint;
+import org.akvo.flow.presentation.geoshape.create.entities.ViewFeatures;
 
 import java.util.ArrayList;
 import java.util.List;
